+++ src/app/layout.tsx
@@ -1,6 +1,9 @@
 import type { Metadata, Viewport } from 'next';
 import localFont from 'next/font/local';
 import Script from 'next/script';
+import { Suspense } from 'react';
+import AuthExpiredRedirect from '@/features/auth/providers/AuthExpiredRedirect';
+import AuthSessionGuard from '@/features/auth/providers/AuthSessionGuard';
 import AuthSessionSync from '@/features/auth/providers/AuthSessionSync';
 import AuthTokenRefreshProvider from '@/features/auth/providers/AuthTokenRefreshProvider';
 import { COMMON_OPEN_GRAPH } from '@/shared/constants/metadata';
@@ -84,9 +87,13 @@ export default function RootLayout({
           type="application/ld+json"
           dangerouslySetInnerHTML={{ __html: JSON.stringify(STRUCTURED_DATA) }}
         />
+        <Suspense fallback={null}>
+          <AuthExpiredRedirect />
+        </Suspense>
         <QueryProvider>
           <AuthTokenRefreshProvider>
             <AuthSessionSync />
+            <AuthSessionGuard />
             {children}
             <Script
               src="//t1.kakaocdn.net/mapjsapi/bundle/postcode/prod/postcode.v2.js"
