+++ src/features/auth/guards/GuestGuard.tsx
@@ -2,8 +2,10 @@
 
 import { useRouter } from 'next/navigation';
 import type { PropsWithChildren } from 'react';
-import { useEffect } from 'react';
+import { useEffect, useRef } from 'react';
 import { useShallow } from 'zustand/shallow';
+import { getMe } from '@/features/user/apis';
+import { ApiError } from '@/shared/apis/apiError';
 import { useUserStore } from '@/shared/stores/userStore';
 
 /**
@@ -19,12 +21,33 @@ export default function GuestGuard({ children }: PropsWithChildren) {
       hasHydrated: state.hasHydrated,
     }))
   );
+  const clearSession = useUserStore((state) => state.clearSession);
+  const didCheckRef = useRef(false);
 
   useEffect(() => {
     if (!hasHydrated) return;
     if (!user) return;
-    router.replace('/');
-  }, [hasHydrated, router, user]);
+
+    // NOTE: store에 user가 남아있지만 쿠키 세션이 만료/삭제된 경우를 방지합니다.
+    if (didCheckRef.current) return;
+    didCheckRef.current = true;
+
+    const check = async () => {
+      try {
+        await getMe();
+        router.replace('/');
+      } catch (error) {
+        if (error instanceof ApiError && error.status === 401) {
+          clearSession('expired');
+          return;
+        }
+        // 네트워크 등 기타 오류는 보수적으로 로그인 화면을 허용합니다.
+        clearSession('expired');
+      }
+    };
+
+    void check();
+  }, [clearSession, hasHydrated, router, user]);
 
   if (!hasHydrated) return null;
   if (user) return null;
